% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaBEHERgui-package.R
\docType{package}
\name{AquaBEHERgui-package}
\alias{AquaBEHERgui}
\alias{AquaBEHERgui-package}
\title{AquaBEHERgui: A shiny GUI app for estimation of wet-season calendar parameters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A shiny GUI app used to compute and visualize soil-water balance and wet-season calendar parameters. The app computes and integrates daily potential evapotranspiration (PET) into a daily soil-water balance model to estimate the calendar of a wet-season for agricultural applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RobelTakele/AquaBEHERgui}
  \item \url{https://robeltakele.github.io/AquaBEHERgui/}
  \item Report bugs at \url{https://github.com/RobelTakele/AquaBEHERgui/issues}
}

}
\keyword{internal}
